Autoinstall the extension
Installing the extension from the chrome store will cause it to be
installed on every chrome browser your account is sync'd with.  Adding
a system policy that auto installs the extension will install it when
browserpass is available on the system but will leave it off of sync'd
devices that don't want it installed.  Browserpass uniquely has an AUR
package and therefore is one of the few google-chrome <--> pass
bridges that can incorporate this feature.