Use background page for filling in forms
This commit fixes the issue of Browspass canceling the filling in when the popup loose focus.
By using a background page Browserpass can continue to retrieve the password even when the popup is closed.
This can happen when the password store isn't unlocked and GPG has to ask for a password.