Fix username bug on windows (#131)
On windows, browserpass wouldn't detect the username automatically
from the password entry name. This was due to the fact that windows
uses backslashes (\) as path seperators, while Linux (read: everybody
else in the entire world) uses forward slashes (/)

This fixes the guessUsername function to always convert the path to be
slash-delimited.